requirements.txt
# Core Discord Bot Dependencies
discord.py>=2.4.0
aiohttp==3.9.1
python-dotenv==1.0.0
requests==2.31.0

# College Football Data API
cfbd>=5.0.0

# Dashboard Dependencies
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
jinja2>=3.1.2
httpx>=0.26.0
python-multipart>=0.0.6
itsdangerous>=2.1.2

# Optional AI Integration
openai==1.6.1
anthropic==0.7.8

# Optional Google Docs Integration
google-api-python-client==2.110.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0

# Development Dependencies (optional)
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
